+++ components/core/pipe/humanize/humanize-duration.pipe.ts
