+++ components/core/pipe/public-api.ts
@@ -6,6 +6,6 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-export * from './nz-pipe.module';
-export * from './nz-time-range.pipe';
-export * from './nz-css-unit.pipe';
+export * from './pipe.module';
+export * from './time-range.pipe';
+export * from './css-unit.pipe';
